backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
requests==2.31.0
python-dotenv==1.0.0
pydantic==2.5.0
SQLAlchemy==2.0.23
greenlet==3.0.3
pymysql==1.1.0
passlib==1.7.4
bcrypt==4.0.1
typing-extensions>=4.7.1,<5.0.0
